+++ packages/vsx-registry/src/browser/vsx-extensions-contribution.ts
@@ -34,11 +34,6 @@ import { BUILTIN_QUERY, INSTALLED_QUERY, RECOMMENDED_QUERY } from './vsx-extensi
 import { IGNORE_RECOMMENDATIONS_ID } from './recommended-extensions/recommended-extensions-preference-contribution';
 import { VSXExtensionsCommands } from './vsx-extension-commands';
 
-/**
- * @deprecated since 1.17.0. - Moved to `vsx-extension-commands.ts` to avoid circular dependencies. Import from there, instead.
- */
-export { VSXExtensionsCommands };
-
 @injectable()
 export class VSXExtensionsContribution extends AbstractViewContribution<VSXExtensionsViewContainer>
     implements ColorContribution, FrontendApplicationContribution {