+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - [output] removed `setVisibility` from `OutputChannelManager` []()
 - [output] removed deprecated const `OUTPUT_WIDGET_KIND` []()
 - [workspace] removed deprecated `getDefaultWorkspacePath` []()
+- [vsx-registry] removed deprecated `VSXExtensionsCommands` re-export []()
 
 ## v1.25.0 - 4/28/2022